Do not inject OOMs in debug diagnostic code paths
If assert and co are about to collect the stacktrace and print diagnostics,
allow them to use the C++ heap even if the test asked for std::bad_alloc being
thrown on the next allocation, because the debug diagnostics are more important.
Implement this by splitting test_heap.hpp out of heap.hpp, including it in
assert.hpp, & disabling OOM injections in all the relevant functions.
