+++ heap.hpp
@@ -1,4 +1,4 @@
-// Copyright 2020-2022 Laurynas Biveinis
+// Copyright 2020-2023 Laurynas Biveinis
 #ifndef UNODB_DETAIL_HEAP_HPP
 #define UNODB_DETAIL_HEAP_HPP
 
@@ -6,10 +6,8 @@
 
 #include <algorithm>
 #ifndef NDEBUG
-#include <atomic>
 #include <cerrno>
 #endif
-#include <cstdint>
 #include <cstdlib>
 #include <new>
 
@@ -18,51 +16,7 @@
 #endif
 
 #include "assert.hpp"
-
-namespace unodb::test {
-
-class allocation_failure_injector final {
- public:
-  static void reset() noexcept {
-#ifndef NDEBUG
-    fail_on_nth_allocation_.store(0, std::memory_order_relaxed);
-    allocation_counter.store(0, std::memory_order_release);
-#endif
-  }
-
-  static void fail_on_nth_allocation(
-      std::uint64_t n UNODB_DETAIL_USED_IN_DEBUG) noexcept {
-#ifndef NDEBUG
-    fail_on_nth_allocation_.store(n, std::memory_order_release);
-#endif
-  }
-
-#ifndef NDEBUG
-
-  UNODB_DETAIL_DISABLE_GCC_WARNING("-Wanalyzer-malloc-leak")
-
-  static void maybe_fail() {
-    const auto fail_counter =
-        fail_on_nth_allocation_.load(std::memory_order_acquire);
-    if (UNODB_DETAIL_UNLIKELY(fail_counter != 0) &&
-        (allocation_counter.fetch_add(1, std::memory_order_relaxed) >=
-         fail_counter - 1)) {
-      throw std::bad_alloc{};
-    }
-  }
-
-  UNODB_DETAIL_RESTORE_GCC_WARNINGS()
-
- private:
-  // NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
-  inline static std::atomic<std::uint64_t> allocation_counter{0};
-  // NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
-  inline static std::atomic<std::uint64_t> fail_on_nth_allocation_{0};
-
-#endif  // #ifndef NDEBUG
-};
-
-}  // namespace unodb::test
+#include "test_heap.hpp"
 
 namespace unodb::detail {
