+++ test/test_utils.hpp
@@ -1,10 +1,10 @@
-// Copyright 2022 Laurynas Biveinis
+// Copyright 2022-2023 Laurynas Biveinis
 #ifndef UNODB_DETAIL_TEST_UTILS_HPP
 #define UNODB_DETAIL_TEST_UTILS_HPP
 
 #include "global.hpp"
 
-#include "heap.hpp"
+#include "test_heap.hpp"
 
 namespace unodb::test {
 
@@ -15,14 +15,14 @@ template <typename TestAction>
 std::invoke_result_t<TestAction> must_not_allocate(
     TestAction test_action) noexcept(noexcept(test_action())) {
   if constexpr (!std::is_void_v<std::invoke_result_t<TestAction>>) {
-    unodb::test::allocation_failure_injector::fail_on_nth_allocation(1);
+    UNODB_DETAIL_FAIL_ON_NTH_ALLOCATION(1);
     auto result = test_action();
-    unodb::test::allocation_failure_injector::reset();
+    UNODB_DETAIL_RESET_ALLOCATION_FAILURE_INJECTOR();
     return result;
   } else {
-    unodb::test::allocation_failure_injector::fail_on_nth_allocation(1);
+    UNODB_DETAIL_FAIL_ON_NTH_ALLOCATION(1);
     test_action();
-    unodb::test::allocation_failure_injector::reset();
+    UNODB_DETAIL_RESET_ALLOCATION_FAILURE_INJECTOR();
   }
 }
 UNODB_DETAIL_RESTORE_MSVC_WARNINGS()
